Colocando Swagger

+++ larmoby/src/main/java/com/example/larmoby/model/Categoria.java
@@ -8,8 +8,7 @@ public class Categoria {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-
-    private int id_categoria;
+    private Long id_categoria;
     private String nome;
 
     // Construtores
@@ -21,11 +20,11 @@ public class Categoria {
 
     // Getters e Setters
 
-    public int getId_categoria() {
+    public Long getId_categoria() {
         return id_categoria;
     }
 
-    public void setId_categoria(int id_categoria) {
+    public void setId_categoria(Long id_categoria) {
         this.id_categoria = id_categoria;
     }
